AST when error for identifier, string, and list
Error actions were added in order to construct partial results within
the AST. This AST is accessible in the buffers hash with the same name
as when completing without error.

The error state does not capture error metadata.

Added spec tests for identifier, string, and list using property tests.
Rantly was added in to the gem.deps.rb file.

refs #3