define FUNCTION as word characters; use in term
Originally I tried being liberal with the character class for FUNCTION,
but it resulted in hung ragel compilation. This is likely due to
ambiguities between two machines, but it's not worth figuring out for
this feature.

Now FUNCTION recognizes a-z, A-Z, 0-9, and _.

Term uses the function ragel machine to accumulate function AST.

closes #64