aggregate expanded line over multiple block calls
Analyzing stacktraces indicated many threads were being created with
calls to Enumerator.next on JRuby. These threads stayed did not complete
and ultimiately resulted in an out of memory error.

The solution employed is to process all lines yielded to the block but
expand, in a stateful manner, when a line continuator is encountered.

JRuby bug: jruby/jruby#2577

remove line_continuator mixin; orphaned

closes OpenBEL/bel.rb#126