improve namespace value lookup
Namespace value lookup is performed against RDF namespace concepts
using the dcterms:identifier, skos:prefLabel, dcterms:title, and
skos:altLabel values. The approach was to select the first match,
but this is incorrect when multiple matches result.

The improvement is to prefer the first namespace value that matched
skos:prefLabel, dcterms:identifier, or dcterms:title. If one does
not exist, such as when two altLabel hits result, we
deterministically pick the first one.

This is a general improvement which affects parsing to the
expression model where namespace value resolution is performed.

closes #142